+++ NEWS.md
@@ -1,4 +1,5 @@
 ## 1.0.0 (IN PROGRESS)
+* ERM-1643: Manage tenant widget definitions for different applications
 * ERM-1580 Added weight to WidgetInstances on dashboard
 * ERM-1579 WidgetDef tweaks, added UUID valueType
 * ERM-1562 Formatting of bootstrapped data, added licenses definition and "Enum" valueType